requirements.txt
cx-Freeze==5.0.2
olefile==0.44
Pillow==4.1.1
pkg-resources==0.0.0
